Fix searching in table editor clears tabs
