+++ apps/studio/pages/project/[ref]/editor/[id].tsx
@@ -3,8 +3,8 @@ import { useCallback, useEffect } from 'react'
 
 import { useParams } from 'common'
 import { useIsTableEditorTabsEnabled } from 'components/interfaces/App/FeaturePreview/FeaturePreviewContext'
-import { TableGridEditor } from 'components/interfaces/TableGridEditor'
 import DeleteConfirmationDialogs from 'components/interfaces/TableGridEditor/DeleteConfirmationDialogs'
+import { TableGridEditor } from 'components/interfaces/TableGridEditor/TableGridEditor'
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import { EditorBaseLayout } from 'components/layouts/editors/EditorBaseLayout'
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
@@ -21,11 +21,14 @@ const TableEditorPage: NextPageWithLayout = () => {
   const store = getTabsStore(projectRef)
 
   const { project } = useProjectContext()
-  const { data: selectedTable, isLoading } = useTableEditorQuery({
-    projectRef: project?.ref,
-    connectionString: project?.connectionString,
-    id,
-  })
+  const { data: selectedTable, isLoading } = useTableEditorQuery(
+    {
+      projectRef: project?.ref,
+      connectionString: project?.connectionString,
+      id,
+    },
+    { retry: false }
+  )
 
   const router = useRouter()
   const onAfterDeleteTable = useCallback(
